+++ src/keystore/index.js
@@ -4,3 +4,4 @@
 export * from './encryption.js';
 export { default } from './encryption.js';
 export * from './provider.js';
+export * from './session-keystore.js';